teste arrumado get e delete

+++ models/productsModel.js
@@ -12,37 +12,38 @@ const getById = async (id) => {
       .then((db) => db.collection('products').findOne({ _id: ObjectId(id) }));
 };
 
-const findByName = async (name) => {
-    connect()
-      .then((db) => db.collection('products').findOne({ name }));
-};
+// const findByName = async (name) => {
+//     connect()
+//       .then((db) => db.collection('products').findOne({ name }));
+// };
 
-const create = async (name, quantity) => {
-    const db = await connect();
-    console.log('db ------- productModel', db);
-    const product = await db.collection('products').insertOne({ name, quantity });
-    console.log('product ------- productModel', product);
-    return { _id: product.insertedId, name, quantity };
-};
+// const create = async (name, quantity) => {
+//     const db = await connect();
+//     console.log('db ------- productModel', db);
+//     const product = await db.collection('products').insertOne({ name, quantity });
+//     console.log('product ------- productModel', product);
+//     return { _id: product.insertedId, name, quantity };
+// };
 
-const updateById = async (name, quantity, id) => {
-    const db = await connect();
-    await db.collection('products')
-      .updateOne({ _id: ObjectId(id) }, { $set: { name, quantity } });
-    return { id, name, quantity };
-};
+// const updateById = async (name, quantity, id) => {
+//     const db = await connect();
+//     await db.collection('products')
+//       .updateOne({ _id: ObjectId(id) }, { $set: { name, quantity } });
+//     return { id, name, quantity };
+// };
 
 const remove = async (id) => {
+    if (!ObjectId.isValid(id)) return null;
     const db = await connect();
     const result = await db.collection('products').deleteOne({ _id: ObjectId(id) });
     return { _id: result.insertId, name: result.name, quantity: result.quantity };
 };
 
 module.exports = { 
-    create,
-    findByName,
+    // create,
+    // findByName,
     getAll,
     getById,
-    updateById,
+    // updateById,
     remove,
 };
